refactor: 북마크와 검색 히스토리 도메인 구조 분리 및 패키지 재구성
