+++ src/main/java/leets/bookmark/domain/searchhistory/application/dto/response/SearchHistoryResponse.java
@@ -1,4 +1,4 @@
-package leets.bookmark.domain.server.application.dto.response;
+package leets.bookmark.domain.searchhistory.application.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
