+++ src/main/java/leets/bookmark/domain/searchhistory/domain/entity/SearchHistory.java
@@ -1,4 +1,4 @@
-package leets.bookmark.domain.server.domain.entity;
+package leets.bookmark.domain.searchhistory.domain.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
